requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
executing==1.2.0
idna==3.4
inflection==0.5.1
ipython==8.7.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.6
pickleshare==0.7.5
prompt-toolkit==3.0.37
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6